013_stock-alert-bot/requirements.txt
yfinance
python-dotenv
discord.py
